Compilation fixes for Arch Linux

+++ radio/src/gui/212x64/lcd.cpp
@@ -532,7 +532,7 @@ void drawTimer(coord_t x, coord_t y, putstime_t tme, LcdFlags att, LcdFlags att2
     tme = -tme;
   }
 
-  qr = div(tme, 60);
+  qr = div((int)tme, 60);
 
 #define separator ':'
   if (att & TIMEHOUR) {
