+++ radio/src/translations/tts_de.cpp
@@ -129,7 +129,7 @@ I18N_PLAY_FUNCTION(de, playNumber, getvalue_t number, uint8_t unit, uint8_t att)
 #else
     // we assume that we are PREC1
 #endif
-    div_t qr = div(number, 10);
+    div_t qr = div((int)number, 10);
     if (qr.rem > 0) {
       PUSH_NUMBER_PROMPT(qr.quot);
       PUSH_NUMBER_PROMPT(DE_PROMPT_COMMA);
